requirements.txt
numpy>=2.3.2
pandas>=2.3.1
scikit-learn>=1.7.1
scipy>=1.16.1
matplotlib>=3.10.3
seaborn>=0.13.2
torch>=2.9.1
sentence-transformers>=5.1.2
faiss-cpu>=1.13.1
transformers>=4.57.3